\name{mzmatch.ipeak.db.CreateCycDB}
\alias{mzmatch.ipeak.db.CreateCycDB}
\title{mzmatch.ipeak.db.CreateCycDB}

\description{
General description of the function of the tool.
}

\usage{
mzmatch.ipeak.db.CreateCycDB(JHeapSize=1425, i, o, h, v)
}

\arguments{
\item{JHeapSize}{
	Amount of RAM memory in megabytes, which should be reserved by the Java virtual machine. The default value is 1425 megabytes.}
\item{i}{
Descriptive text for the option.
}
\item{o}{
Descriptive text for the option.
}
\item{h}{
When this is set, the help is shown.
}
\item{v}{
When this is set, the progress is shown on the standard output.
}

}

\details{
General description of the function of the tool.
}

\value{
This function returns no value.
}

\references{
0. PeakML/mzMatch - a file format, Java library, R library and tool-chain for mass spectrometry data analysis. In preparation.


}

\author{
RA Scheltema (r.a.scheltema@rug.nl)
}

\seealso{
\code{\link{PeakML.xcms.read}}
}

\keyword{mzMatch}
\keyword{peakML}
\keyword{java tools}
