\name{mzmatch.ipeak.filter.FilterCluster}
\alias{mzmatch.ipeak.filter.FilterCluster}
\title{mzmatch.ipeak.filter.FilterCluster}

\description{
Filter cluster
}

\usage{
mzmatch.ipeak.filter.FilterCluster(JHeapSize=1425, i, o, rejected, formula, h, v)
}

\arguments{
\item{JHeapSize}{
	Amount of RAM memory in megabytes, which should be reserved by the Java virtual machine. The default value is 1425 megabytes.}
\item{i}{

}
\item{o}{

}
\item{rejected}{

}
\item{formula}{

}
\item{h}{
When this is set, the help is shown.
}
\item{v}{
When this is set, the progress is shown on the standard output.
}

}

\details{
Filter cluster
}

\value{
This function returns no value.
}

\references{
0. PeakML/mzMatch - a file format, Java library, R library and tool-chain for mass spectrometry data analysis. In preparation.


}

\author{
RA Scheltema (r.a.scheltema@rug.nl)
}

\seealso{
\code{\link{PeakML.xcms.read}}
}

\keyword{mzMatch}
\keyword{peakML}
\keyword{java tools}
