\name{mzmatch.ipeak.normalisation.RelativeIntensityNormalize}
\alias{mzmatch.ipeak.normalisation.RelativeIntensityNormalize}
\title{mzmatch.ipeak.normalisation.RelativeIntensityNormalize}

\description{
Normalise input matrix according relative intensity.
}

\usage{
mzmatch.ipeak.normalisation.RelativeIntensityNormalize(JHeapSize=1425, i, o, sets, v, h)
}

\arguments{
\item{JHeapSize}{
	Amount of RAM memory in megabytes, which should be reserved by the Java virtual machine. The default value is 1425 megabytes.}
\item{i}{

}
\item{o}{

}
\item{sets}{
for each set a reference
}
\item{v}{
When this is set, the progress is shown on the standard output.
}
\item{h}{
When this is set, the help is shown.
}

}

\details{
Normalise input matrix according relative intensity.
}

\value{
This function returns no value.
}

\references{
0. PeakML/mzMatch - a file format, Java library, R library and tool-chain for mass spectrometry data analysis. In preparation.


}

\author{
RA Scheltema (r.a.scheltema@rug.nl)
}

\seealso{
\code{\link{PeakML.xcms.read}}
}

\keyword{mzMatch}
\keyword{peakML}
\keyword{java tools}
