\name{mzmatch.R-package}
\alias{mzmatch.R-package}
\alias{mzmatch.R}
\docType{package}
\title{
mzmatch and PeakML integration for XCMS and R
}
\description{
mzMatch is a modular, open source and platform independent data processing pipeline for metabolomics LC/MS data written in the Java language. It was designed to provide small tools for the common processing tasks for LC/MS data. The mzMatch environment was based entirely on the PeakML file format and core library, which provides a common framework for all the tools.
With the introduction of the PeakML file format, the tools can share their data output with the other software. This means that we can integrate with other tools, for which we provide a first implementation integration mzMatch with XCMS an R environment. This package provides integration between the mzMatch/PeakML Java library (and its file formats) and the XCMS R library. Content of the mzmatch.R library is extended with several R functions for visualisation and statistical analysis of LC/MS data sets.
License: GNU General Public License (GPL)
}
\details{
\tabular{ll}{
Package: \tab mzmatch.R\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-10-20\cr
License: \tab GNU General Public License (GPL)\cr
LazyLoad: \tab yes\cr
}
}
\author{
RA Scheltema (r.a.scheltema@rug.nl), A Jankevics (a.jankevics@rug.nl)

Maintainer: Andris Jankevics <a.jankevics@rug.nl>
}
\references{
PeakML/mzMatch - a file format, Java library, R library and tool-chain for mass spectrometry data analysis. In preparation.
}
\keyword{mzmatch.R}
\keyword{peakML}
\keyword{mzMatch}
