\name{PeakML.xcms.read}
\alias{PeakML.xcms.read}
\title{PeakML.xcms.read}
\description{
  This function reads the contents of the given PeakML file. This file format
  supports the mass traces (features) and the sets-structure utilized by XCMS and as such provides a
  fully reproducible representation of the contents of an XCMS set structure.
}
\usage{
  PeakML.xcms.read(filename, ionisation = "detect", Rawpath = NULL, annotations = FALSE, version.1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{filename}{
    The name of the PeakML file to load. This file needs to exist, otherwise the function will fail.
  }
  \item{ionisation}{
    This parameter is obsolete.
  }
  \item{Rawpath}{
    The path to the raw data, which is needed if the parameter getretentiontimefromraw is set to TRUE and the path
	stored in the PeakML file .
  }
  \item{annotations}{
    When this is set to TRUE, the function also loads the stored annotations from the PeakML file.
  }
  \item{version.1}{
	 For bacwkard compability it is possible to use mzmatch V1. If option version.1 is set to "FALSE", it will load the latest mzmatch version. PeakML files generated with mzmatch V.1 are not compatiable with V.2 and above and vice versa.
	}
}
\details{
  The PeakML file format defines an open and extensible file format for storing extracted hyphenated mass spectrometry
  data. As such, the format will enable users to store intermediate data in a structured manner, but more importantly,
  to exchange data between different data analysis software. This will give complete flexible control to the users of
  data analysis software to cherry pick those components best suited to their needs.
  
  The PeakML.xcms.read function extends the XCMS package with read functionality for importing data generated by other
  applications. For example, the results of from actions in the user interface environment PeakML Viewer can be loaded.
  
  The function allows the raw-path entries in the PeakML file to be overridden with a different path, to enable usage
  on a different computer where the files are stored in a different path. The PeakML file format does not support
  the storage of uncorrected retention times (which is featured in the XCMS set structure). Set the parameter
  rtfromraw to TRUE to retrieve the original retention times from the raw files.
  
  The PeakML file format supports annotations, which can be loaded into the XCMS set structure by setting the
  parameter annotations to true.
}
\value{
  This function returns an XCMS set structure when it succeeds, NULL otherwise. When the parameter annotations
  has been set to TRUE a list is returned with the first element the XCMS set and the second the annotations.
}
\references{
  PeakML/mzMatch - a file format, Java library, R library and tool-chain for mass spectrometry data analysis. In preparation.
}
\author{
  Andris Jankevics (a.jankevics@rug.nl), Richard Scheltema (r.a.scheltema@rug.nl)
}

\seealso{
  \code{\link{PeakML.xcms.write}}
}

\keyword{ mzmatch.R }
\keyword{ XCMS }
